+++ CHANGES.md
@@ -5,6 +5,7 @@ Unreleased
 * Add `Re.exec_partial_detailed` to allow resuming searches from partial inputs
   (#219)
 * Add support for `DOTALL` flag in `Re.Pcre.regexp` (#225)
+* Add support for named groups (#223)
 
 1.10.4 (27-Apr-2022)
 --------------------